Added kiwix-serve external link custom page
Should eventually be included in kiwix-serve (https://github.com/kiwix/kiwix-tools/issues/711)
but dashboard is best home for it ATM as its always included.

As its independent, it doesn't rely on any dashboard asset or anything

Uses a template only so it can provide a <base /> with correct FQDN as this is meant to be reversed
and not accessed directly.

Also requires CORS header as page calls for fonts.