Fixed incorrect CSS when using image
In development, sources files are templates ;
when running in image, source files are rendered templates at different
locations.
tailwindcss was thus not able to find source file and could not compile
the final CSS as it should be.
Also, the tailwind config file (referencing source files and other)
was not to be found.